Graduates the "stand alone" ann tooling
As we were working on this project, we needed to run targeted
experiments outside of the framework that ann-benchmark provides. We
started calling the work in the aerospike directory "stand alone ann".
That work has long since become a distinct project, and this commit
removes that work from this repository.

The tool has been moved to

https://github.com/aerospike-community/avs-benchmark-ann

and will continue to evolve separately from this repository.